Update post metadata saving to use new URL parser
This started life as a task to include note ID in the post metadata when it was keyed off a note, since the document slug is the same on notes and documents, and thus two resources from the same doc could cause a ID collision.

Then I realized that we already only save a single piece of metadata per post anyway (see how `update_post_meta()` is outside the `foreach()`?) and so this was unnecessary.

But I'd already done the work to abstract out `parse_dc_url()` and clean up some things, so.

Affects #22