Disable toolbar button until resolved
Toolbar button has been broken for a while, since at least 3.9.1, because of an update to the version of TinyMCE included in WP. It needs some signifiant fixing, and could stand to have it fleshed out with additional attribute options anyway. Disabling until we have time for the full solution.

Affects #23