Append note ID to postmeta key for notes
Docs and notes have the same document slug, which resulted in the same postmeta key. When you had a wide document followed by a normal note, this resulted in the postmeta not properly reporting a `wide_asset`. Append note ID to its meta key so it can be distinguished from docs.

Affects #22